@@ -12,30 +12,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "16.2.3",
-    "@angular/cdk": "16.2.12",
-    "@angular/common": "16.2.3",
-    "@angular/compiler": "16.2.3",
-    "@angular/core": "16.2.3",
-    "@angular/forms": "16.2.3",
-    "@angular/platform-browser": "16.2.3",
-    "@angular/platform-browser-dynamic": "16.2.3",
-    "@angular/router": "16.2.3",
+    "@angular/animations": "17.0.6",
+    "@angular/cdk": "17.0.3",
+    "@angular/common": "17.0.6",
+    "@angular/compiler": "17.0.6",
+    "@angular/core": "17.0.6",
+    "@angular/forms": "17.0.6",
+    "@angular/platform-browser": "17.0.6",
+    "@angular/platform-browser-dynamic": "17.0.6",
+    "@angular/router": "17.0.6",
     "@types/file-saver": "2.0.7",
     "angular-file-saver": "1.1.3",
     "chart.js": "4.4.1",
     "primeflex": "3.3.1",
     "primeicons": "6.0.1",
-    "primeng": "16.9.1",
+    "primeng": "17.0.0",
     "rxjs": "7.8.1",
     "tslib": "2.6.2",
-    "zone.js": "0.13.3"
+    "zone.js": "0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "16.2.9",
+    "@angular-devkit/build-angular": "17.0.6",
     "@angular/cli": "17.0.6",
-    "@angular/compiler-cli": "16.2.3",
+    "@angular/compiler-cli": "17.0.6",
     "sync-pom-version-to-package": "1.6.1",
-    "typescript": "5.1.6"
+    "typescript": "5.2.2"
   }
 }
