@@ -1,13 +1,13 @@
 import {Component, OnInit} from '@angular/core';
 import {RouterLink} from '@angular/router';
 import {Breadcrumb, Principal, Url, User} from "@naikan/shared";
-import {NgForOf, NgIf} from "@angular/common";
+
 import {TagModule} from "primeng/tag";
 
 @Component({
   templateUrl: './profile.component.html',
   standalone: true,
-  imports: [RouterLink, Breadcrumb, NgIf, Url, NgForOf, TagModule]
+  imports: [RouterLink, Breadcrumb, Url, TagModule]
 })
 export class ProfileComponent implements OnInit {
 
